### 2018-02-25 b9cf105
Added bib file for joss paper
