### 2018-02-25 #595 @motib
Hello,

1. When you say "type @whedon commands", where do I type it??

2. I looked at the list of reviewers. No one is actually in the field of the application of logic to computer science (verification, SAT solving, logic programming, etc.), but krother and hainesr might be appropriate
because of their interest in education.

3. Sorry about the bib file, but the instructions didn't say where to put it.
It is in the directory "docs".
Should I have used "docs/learnsat.bib" as the file name?
Should I have uploaded it to the top level directory with "paper.md".
I am guess that I should have uploaded there, so I have now done so.

I am totally new to Markdown and have limited experience with Git,
so I apologize for any problems and will be grateful for your help.

@motib
