### 2018-02-26 327ad0d
Modifications after first JOSS compilation
