### 2018-03-12 #595 @danielskatz
@shnarazk, @niklaseen, @alanminko - would you be willing to review this submission to JOSS?
