### 2018-03-13 #595 @motib
Hi Daniel, Looking at your suggestions, @shnarazk works in SAT solving, the same area as my submission, and would be the most qualified to review it.
