### 2018-03-16 #595 @danielskatz
@biotomas - would you be willing to review this submission for JOSS?
