### 2018-03-21 #639 @whedon
[REVIEW]: LearnSAT: A SAT Solver for Education
**Submitting author:** @motib (<a href="http://orcid.org/0000-0002-2091-5107">Mordechai Ben-Ari</a>)
**Repository:** <a href="https://github.com/motib/LearnSAT" target ="_blank">https://github.com/motib/LearnSAT</a>
**Version:** V2.0
**Reviewer1:** @audemard
**Reviewer2:** @biotomas
