### 2018-03-21 #639 @danielskatz
@audemard & @biotomas - this review is now open.
See instructions above.

Please use this issue to track any questions/concerns you have, even if you actually create new issues on focused details in https://github.com/motib/LearnSAT

Any questions on the review, please ask.
