### 2018-03-21 #639 @motib
@audemard , @biotomas,

Thank you for agreeing to review my submission.

I am leaving today for a 12-day vacation (really, I'm not taking a laptop!).
While I will be able to respond to general queries,
addressing detailed questions about the software
will have to wait until the beginning of April.
