### 2018-03-26 #639 @danielskatz
@audemard @biotomas - I'm hoping that you both can start your reviews soon.

In case it's not clear, your job is basically to go through the checklist items, making sure that all can be checked.  To do this, you will also have to examine the [paper](https://github.com/openjournals/joss-papers/blob/joss.00639/joss.00639/10.21105.joss.00639.pdf) and the contents of the [code repository](https://github.com/motib/LearnSAT).

If you can check off an item, please do.  If you cannot, please either discuss it here if it is brief, or open an issue in the [code repository](https://github.com/motib/LearnSAT) and put a comment here pointing to that issue.

Thanks.
