### 2018-04-03 #639 @biotomas
Hi, so I left two checkboxes unchecked because I found no DOIs in references and also no community guidelines.
