### 2018-04-03 #639 @danielskatz
@motib - In response to the comments by @biotomas:

1. It looks like the handbook chapters have DOIs (http://ebooks.iospress.nl/volume/handbook-of-satisfiability) so you should be able to add at least that one DOI.  I don't think there are DOIs for the two books, but if they have similar identifiers, such as ISBNs, please add them.

2. While looking at https://github.com/motib/LearnSAT/blob/master/learnsat.bib, I also notice you have quite a few entries in this file that are not actually cited in the paper.md file and thus are not references. I don't know if you want to cite any more of the bib entries.

3. I also see that there is no References header in the paper.md file.   I've created https://github.com/motib/LearnSAT/pull/1 to fix this.

4. Please also address the lack of community guidelines.
