### 2018-04-04 #639 @motib
My thanks to @biotomas for his comments.

1. I have chased down DOIs and ISBNs for all the references and will update the documents.

2. The file "learnsat.bib" contains the bibtex entries for _all_ three of the documents included in the archive as well as those in paper.md. Since latex is smart enough to extract only the references cited for each document, I don't see this as a problem. To the contrary, if I had four separate files, there would be a risk of inconsistency.

3. I am rather new to git. Would it be correct for me simply to click "merge"?

4. I'm sorry for not noticing this. The submission instructions don't specify where this information should be included. In the "readme"? In a separate file? If so, is there a convention for naming it?

@danielskatz: please clarify points 3 and 4 and I'll update the submission.
