### 2018-04-04 #639 @danielskatz
@motib -

> 3. I am rather new to git. Would it be correct for me simply to click "merge"?

yes, though of course you should first check the changed file(s) to make sure you want to merge it.

> 4. I'm sorry for not noticing this. The submission instructions don't specify where this information should be included. In the "readme"? In a separate file? If so, is there a convention for naming it?

This is commonly in both the readme and the documentation, but there is no requirement - it should be where you think it will be the most useful to potential contributors.
