### 2018-04-05 4d7507b
Add files via upload
Added references to paper.md and added community-guidelines.md.
