### 2018-04-05 2ca313d
Delete learnsat.bib
Correcting mistake - learnsat.bib should be in parent directory
