+++ docs/learnsat.bib
@@ -1,122 +1,0 @@
-@unpublished{miller-lang,
-  author={J.E. Miller},
-  title={{Langford's Problem}},
-  note={\url{http://dialectrix.com/langford.html}},
-}
-
-@book{knuth-sat,
- author={D.E. Knuth},
- title={The Art of Computer Programming, Volume 4, Fascicle 6: Satisfiability},
- publisher={Pearson},
- note={ISBN 978-0-1343-9760-3},
- year={2015}
-}
-
-@article{brute,
- author = {Heule, Marijn J. H. and Kullmann, Oliver},
- title = {The Science of Brute Force},
- journal = {Commun. ACM},
- volume = {60},
- number = {8},
- year = {2017},
- note={\url{http://dx.doi.org/10.1145/3107239}},
- pages = {70--79},
-} 
-
-@book{SAT,
-  title={Handbook of Satisfiability},
-  editor={A. Biere and M. Heule and H. Van Maaren and T. Walsh},
-  year=2009,
-  note={ISBN 978-1-60750-376-7},
-  publisher={{IOS} Press},
-  }
-
-@book{mlcs,
-  title={Mathematical Logic for Computer Science (Third Edition)},
-  author={M. Ben-Ari},
-  publisher={Springer},
-  note={ISBN	978-1-4471-4129-7},
-  year=2012,
-}
-
-@inbook{mlm,
-   Author = {Marques-Silva, J. P. and Lynce, I. and Malik, S.},
-   Title = {Conflict-Driven Clause Learning SAT Solvers},
-   Chapter = {4},
-   Pages = {131--153},
-   note={\url{http://dx.doi.org/10.3233/978-1-58603-929-5-131}},
-   Crossref = {SAT}
-}
-
-@inbook{look,
-   author = {Heule, Marijn J. H. and {van Maaren, H.}},
-   Title = {Look-Ahead Based SAT Solvers},
-   Chapter = {5},
-   Pages = {155--184},
-  note={\url{http://dx.doi.org/10.3233/978-1-58603-929-5-155}},
-   Crossref = {SAT}
-}
-
-
-@inbook{bmc,
-   Author = {Biere, A.},
-   Title = {Bounded Model Checking},
-   Chapter = {14},
-   Pages = {457--481},
-   note={\url{http://dx.doi.org/10.3233/978-1-58603-929-5-457}},
-   Crossref = {SAT}
-}
-
-@inbook{cnf,
-   Author = {Prestwich, S.},
-   Title = {CNF Encodings},
-   Chapter = {2},
-   Pages = {75--97},
-   note={\url{http://dx.doi.org/10.3233/978-1-58603-929-5-75}},
-   Crossref = {SAT}
-}
-
-@article{mz,
- author = {Malik, S. and Zhang, L.},
- title = {Boolean satisfiability from theoretical hardness to practical success},
- journal = {Commun. ACM},
- volume = {52},
- number = {8},
- year = {2009},
- note={\url{http://dx.doi.org/10.1145/1536616.1536637}},
- pages = {76--82},
-} 
-
-@inproceedings{ms,
- author = {Marques-Silva, J. P. and Sakallah, K. A.},
- title = {GRASP: A search algorithm for propositional satisfiability},
- journal= {IEEE Transactions on Computers},
- volume=48,
- number=5,
- year = {1999},
- note={\url{http://dx.doi.org/10.1109/12.769433}},
- pages = {506--521},
-}
-
-@article{nadel,
- author = {Nadel, B.A.},
- title = {Representation Selection for Constraint Satisfaction: A Case Study Using n-Queens},
- journal = {IEEE Expert: Intelligent Systems and Their Applications},
- volume = {5},
- issue = {3},
[…]
