+++ learnsat.bib
@@ -8,6 +8,7 @@
  author={D.E. Knuth},
  title={The Art of Computer Programming, Volume 4, Fascicle 6: Satisfiability},
  publisher={Pearson},
+ note={ISBN 978-0-1343-9760-3},
  year={2015}
 }
 
@@ -18,6 +19,7 @@
  volume = {60},
  number = {8},
  year = {2017},
+ note={\url{http://dx.doi.org/10.1145/3107239}},
  pages = {70--79},
 } 
 
@@ -25,6 +27,7 @@
   title={Handbook of Satisfiability},
   editor={A. Biere and M. Heule and H. Van Maaren and T. Walsh},
   year=2009,
+  note={ISBN 978-1-60750-376-7},
   publisher={{IOS} Press},
   }
 
@@ -32,6 +35,7 @@
   title={Mathematical Logic for Computer Science (Third Edition)},
   author={M. Ben-Ari},
   publisher={Springer},
+  note={ISBN	978-1-4471-4129-7},
   year=2012,
 }
 
@@ -40,6 +44,7 @@
    Title = {Conflict-Driven Clause Learning SAT Solvers},
    Chapter = {4},
    Pages = {131--153},
+   note={\url{http://dx.doi.org/10.3233/978-1-58603-929-5-131}},
    Crossref = {SAT}
 }
 
@@ -48,6 +53,7 @@
    Title = {Look-Ahead Based SAT Solvers},
    Chapter = {5},
    Pages = {155--184},
+  note={\url{http://dx.doi.org/10.3233/978-1-58603-929-5-155}},
    Crossref = {SAT}
 }
 
@@ -57,6 +63,7 @@
    Title = {Bounded Model Checking},
    Chapter = {14},
    Pages = {457--481},
+   note={\url{http://dx.doi.org/10.3233/978-1-58603-929-5-457}},
    Crossref = {SAT}
 }
 
@@ -65,6 +72,7 @@
    Title = {CNF Encodings},
    Chapter = {2},
    Pages = {75--97},
+   note={\url{http://dx.doi.org/10.3233/978-1-58603-929-5-75}},
    Crossref = {SAT}
 }
 
@@ -75,16 +83,19 @@
  volume = {52},
  number = {8},
  year = {2009},
+ note={\url{http://dx.doi.org/10.1145/1536616.1536637}},
  pages = {76--82},
 } 
 
 @inproceedings{ms,
  author = {Marques-Silva, J. P. and Sakallah, K. A.},
- title = {GRASP---a new search algorithm for satisfiability},
- booktitle = {Proceedings of the 1996 IEEE/ACM International Conference on Computer-Aided Design},
- series = {ICCAD '96},
- year = {1996},
- pages = {220--227},
+ title = {GRASP: A search algorithm for propositional satisfiability},
+ journal= {IEEE Transactions on Computers},
+ volume=48,
+ number=5,
+ year = {1999},
+ note={\url{http://dx.doi.org/10.1109/12.769433}},
+ pages = {506--521},
 }
 
 @article{nadel,
@@ -95,6 +106,7 @@
  issue = {3},
  month = {June},
  year = {1990},
+ note={\url{http://dx.doi.org/10.1109/64.54670}},
  pages = {16--23},
 } 
 
@@ -105,5 +117,6 @@
  issue_date = {March 2010},
  volume = {1},
  number = {1},
+ note={\url{http://dx.doi.org/10.1145/1721933.1721950}},
  pages = {40--47},
 } 
