### 2018-04-05 #639 @motib
I have made the modifications, but the isbns/dois do not appear on the generated pdf of the "paper". Possibly, s not including the package \url that I use to format the dois. If so, should I remove \url or is there someway to add this package?

Also: @danielskatz I just noticed that you forked the project, but I made the modifications to the master branch. Is that OK?
