### 2018-04-05 #639 @danielskatz
> Also: @danielskatz I just noticed that you forked the project, but I made the modifications to the master branch. Is that OK?

It looks like you made the equivalent change to what was in my branch without merging my branch's PR, which is fine, so I've now deleted the branch and closed the PR.
