### 2018-04-05 #639 @danielskatz
> I have made the modifications, but the isbns/dois do not appear on the generated pdf of the "paper". Possibly, it is not including the package \url that I use to format the dois. If so, should I remove \url or is there someway to add this package?

@arfon, can you help with this?  I would have expected the note data to be pulled from the .bib file into the references.  Perhaps you can suggest how this should be done?
