### 2018-04-05 #639 @arfon
> @arfon, can you help with this? I would have expected the note data to be pulled from the .bib file into the references. Perhaps you can suggest how this should be done?

Sure. I think these should be listed as `doi` fields as well as `url`. So for [this line](https://github.com/motib/LearnSAT/blob/master/learnsat.bib#L22) change it to:

```
doi={10.1145/3107239},
url={http://dx.doi.org/10.1145/3107239}

```
