### 2018-04-05 #639 @danielskatz
And what do you suggest for the ISBNs for the books?  I would have expected the note field to work.
