### 2018-04-06 f4e54eb
Add files via upload
For bibliography change to use hyperref and plainurl for DOIs
