+++ docs/learnsat-ug.tex
@@ -3,6 +3,7 @@
 \usepackage{mathptmx}
 \usepackage{url}
 \usepackage{graphicx}
+\usepackage{hyperref}
 \graphicspath{{images/}}
 
 \newcommand*{\p}[1]{\textup{\texttt{#1}}}
@@ -583,6 +584,6 @@ sorted set of the variables corresponding to the literals.
 
 %\newpage
 
-\bibliographystyle{plain}
+\bibliographystyle{plainurl}
 \bibliography{learnsat}
 \end{document}
