+++ docs/learnsat-overview.tex
@@ -3,6 +3,7 @@
 \usepackage{mathptmx}
 \usepackage{url}
 \usepackage{graphicx}
+\usepackage{hyperref}
 \graphicspath{{images/}}
 
 \newcommand*{\ls}{\textsc{LearnSAT}}
@@ -147,7 +148,7 @@ hole2 :- dpll([
 
 \LaTeX{} and PDF files are provided for the \ls{} User's Guide and Software Documentation and the \ls{} Tutorial.
 
-\bibliographystyle{plain}
+\bibliographystyle{plainurl}
 \bibliography{learnsat}
 
 \newpage
