+++ docs/learnsat-tutorial.tex
@@ -5,6 +5,7 @@
 \usepackage{graphicx,pict2e,epic}
 \usepackage{verbatim}
 \usepackage{wasysym}
+\usepackage{hyperref}
 
 \graphicspath{{images/}}
 
@@ -1934,9 +1935,6 @@ permutations of the order of the variables, there remains an element of
 ``luck'' when you choose an ordering arbitrarily, at random or using
  heuristics.
 
-
-
-
-\bibliographystyle{plain}
+\bibliographystyle{plainurl}
 \bibliography{learnsat}
 \end{document}
