### 2018-04-06 2a821f1
Add files via upload
Change to use "doi=" instead of "note="
