+++ learnsat.bib
@@ -19,7 +19,7 @@
  volume = {60},
  number = {8},
  year = {2017},
- note={\url{http://dx.doi.org/10.1145/3107239}},
+ doi={10.1145/3107239},
  pages = {70--79},
 } 
 
@@ -44,7 +44,7 @@
    Title = {Conflict-Driven Clause Learning SAT Solvers},
    Chapter = {4},
    Pages = {131--153},
-   note={\url{http://dx.doi.org/10.3233/978-1-58603-929-5-131}},
+   doi={10.3233/978-1-58603-929-5-131},
    Crossref = {SAT}
 }
 
@@ -53,7 +53,7 @@
    Title = {Look-Ahead Based SAT Solvers},
    Chapter = {5},
    Pages = {155--184},
-  note={\url{http://dx.doi.org/10.3233/978-1-58603-929-5-155}},
+   doi={10.3233/978-1-58603-929-5-155},
    Crossref = {SAT}
 }
 
@@ -63,7 +63,7 @@
    Title = {Bounded Model Checking},
    Chapter = {14},
    Pages = {457--481},
-   note={\url{http://dx.doi.org/10.3233/978-1-58603-929-5-457}},
+   doi={10.3233/978-1-58603-929-5-457},
    Crossref = {SAT}
 }
 
@@ -72,7 +72,7 @@
    Title = {CNF Encodings},
    Chapter = {2},
    Pages = {75--97},
-   note={\url{http://dx.doi.org/10.3233/978-1-58603-929-5-75}},
+   doi={10.3233/978-1-58603-929-5-75},
    Crossref = {SAT}
 }
 
@@ -83,18 +83,18 @@
  volume = {52},
  number = {8},
  year = {2009},
- note={\url{http://dx.doi.org/10.1145/1536616.1536637}},
+ doi = {10.1145/1536616.1536637},
  pages = {76--82},
 } 
 
-@inproceedings{ms,
+@article{ms,
  author = {Marques-Silva, J. P. and Sakallah, K. A.},
  title = {GRASP: A search algorithm for propositional satisfiability},
  journal= {IEEE Transactions on Computers},
  volume=48,
  number=5,
  year = {1999},
- note={\url{http://dx.doi.org/10.1109/12.769433}},
+ doi={10.1109/12.769433},
  pages = {506--521},
 }
 
@@ -106,7 +106,7 @@
  issue = {3},
  month = {June},
  year = {1990},
- note={\url{http://dx.doi.org/10.1109/64.54670}},
+ doi={10.1109/64.54670},
  pages = {16--23},
 } 
 
@@ -117,6 +117,6 @@
  issue_date = {March 2010},
  volume = {1},
  number = {1},
- note={\url{http://dx.doi.org/10.1145/1721933.1721950}},
+ doi={10.1145/1721933.1721950},
  pages = {40--47},
 } 
