### 2018-04-06 #639 @arfon
> And what do you suggest for the ISBNs for the books? I would have expected the note field to work.

According to this post the `note` field should work https://tex.stackexchange.com/questions/52040/bibtex-show-isbn-number. I'm not sure what bibliography style we're using for JOSS but it's clearly not including `note` fields by default...
