### 2018-04-06 777f49f
Delete community-guidelines.md
Move community guidelines to readme file
