+++ community-guidelines.md
@@ -1,16 +1,0 @@
-# LearnSat - Community Guidelines
-
-**Mordechai (Moti) Ben-Ari**
-
-[http://www.weizmann.ac.il/sci-tea/benari/](http://www.weizmann.ac.il/sci-tea/benari/)
-
----
-
-Thank you for your interest in LearnSAT.
-
-This is my first Open Source / GitHub project, so there may be some glitches until I have more experience!
-
-For now, the guidelines are:
-
-- If you wish to modify the Prolog source code, please fork the project and leave the master branch with my source code.
-- If you have additional SAT examples or tutorial material that you want to add, I can include them in the master branch. Open an issue and / or contact me directly at [moti.ben.ari@gmail.com](mailto:moti.ben.ari@gmail.com).
