+++ README.md
@@ -1,36 +1,0 @@
-# LearnSat
-
-**Mordechai (Moti) Ben-Ari**
-
-[http://www.weizmann.ac.il/sci-tea/benari/](http://www.weizmann.ac.il/sci-tea/benari/)
-
----
-
-LearnSAT is a Prolog program for learning about SAT solving. It
-implements the classic Davis-Putnam-Logemann-Loveland (DPLL) algorithm,
-together will modern extensions of the algorithm: conflict-driven clause
-learning (CDCL), non-chronological backtracking (NCB) and lookahead.
-
-## Licences
-- The software is made available under the GNU General Public License.
-- The documentation is made available under the Creative Commons Attribution-ShareAlike 3.0 License.
-
-## Installation and running
-
-LearnSAT is written in Prolog.
-
-- Install [SWI-Prolog]( http://www.swi-prolog.org/).
-- Unzip the archive `learnsat-N.zip`.
-- To run the LearnSAT, consult the file `dpll.pro` as explained in the User's Guide.
-
-## Directories
-
-- src: The LearnSAT source code.
-- examples: Source code for the examples.
-- docs: Documentation (in LaTeX and PDF).
-
-## Documentation
-
-- `learnsat-overview` An overview of LearnSAT. 
-- `learnsat-ug` User's guide and software documentation.
-- `learnsat-tutorial` Tutorial on SAT solving using LearnSAT.
