### 2018-04-06 b1655e9
Add files via upload
Updated User Guide with links to JOSS article and GitHub repository.
Refer to readme file for information on installation and running and for community guidelines.
