### 2018-04-06 #639 @motib
@arfon: On stackexchange I found a simple way to include DOIs: change the bib style to "plainurl" and include the "hyperref" package. "hyperref" additionally creates internal links in the PDF. Now the documents include "doi" fields with no "http" but the fields are clickable links. On the whedon-generated pdf, the "doi" appears as a link with "http".

There is no simple way include isbn's (without major surgery to the source files that I don't want to do and don't see a need for). The suggestion is to use "note" since books are easy to search for on google or amazon. This works OK for the documents, but not for the whedon-generated pdf. Since that is not under my control, I can't fix it ....

As per @danielskatz's suggestion, I moved the community guidelines to the readme file and referred to it in the User Guide. I took the opportunity to modify the UG by deleting the instruction to download from my personal website; instead, I give links to the JOSS article and to the GitHub repo.
