### 2018-04-06 #639 @arfon
> Is this something you think we can fix, or should we just accept that we can't do this for now?

I think it's worth opening an issue to track this bug but I don't think this can be fixed quickly.
