### 2018-04-06 #639 @danielskatz
ok, I opened https://github.com/openjournals/joss/issues/392
