### 2018-04-26 #639 @motib
@arfon: I have never heard of these before so I have a couple of questions.
I logged into Zenodo (with my GitHub account).

1. It wants me to upload files. Should I update individual files or could I just download a zip of the GitHub repository and upload that? I would prefer the latter since the directories on my computer have additional stuff like my local backups that I don't want to upload.

2. It asks "Did your publisher already assign a DOI to your upload?" Does that mean that I should give them the JOSSS DOI or should I let Zenodo assign one? In the latter case, should I add the JOSS DOI to the "related identifier" field?
Thanks....
