### 2018-04-26 #639 @arfon
> It wants me to upload files. Should I update individual files or could I just download a zip of the GitHub repository and upload that? I would prefer the latter since the directories on my computer have additional stuff like my local backups that I don't want to upload.

@motib - I'd suggest following this guide - it should be pretty simple: https://guides.github.com/activities/citable-code/

> It asks "Did your publisher already assign a DOI to your upload?" Does that mean that I should give them the JOSSS DOI or should I let Zenodo assign one? In the latter case, should I add the JOSS DOI to the "related identifier" field?

We don't assign a DOI for your upload. Let Zenodo assign one (and paste it in this thread).

> In the latter case, should I add the JOSS DOI to the "related identifier" field?

You can do this. Your JOSS DOI will be https://doi.org/10.21105/joss.00639
